refactor createTRansaction method to accept safe settings txs
